+++ docs/source/changelog.rst
@@ -13,7 +13,8 @@ v2.0.3
 - Grouped Point Target and Distributed Spectral Analysis under the main ``spectral-analysis`` command
 - Improved ``InvalidProductType`` exception handling when loading products without having installed their dedicated format plugins
 - Changed ``elevation_profile`` radiometric analysis cli command to ``elevation-profile``
-- Updated radiometric analysis graphs generation inputs to match latest perseo-quality
+- Updated radiometric analysis graphs generation inputs to match latest perseo-quality\
+- Testing: changed KPI statistics tolerance for integration tests validation
 
 **Bug Fixing**